requirements.txt
feedparser
psycopg2-binary
sqlalchemy
bertopic
underthesea
pandas